Moving some files for orginizational clarity.
