+++ app/assets/javascripts/components/browser/browser.jsx
@@ -15,9 +15,9 @@ import * as ReactRedux from 'react-redux';
 
 // Class imports.
 import Magma from '../../magma';
-import Header from '../header';
+import Header from '../general/header';
 import {HelpContainer as Help} from '../help';
-import {TabBarContainer as TabBar} from '../tab_bar';
+import {TabBarContainer as TabBar} from '../general/tab_bar';
 import BrowserTab from './browser_tab';
 
 // Module imports.
